zip with password

+++ app/Http/Controllers/Frontend/FrontendController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use Mail;
 use Input;
 use ZipArchive;
+use DateTime;
+use DateTimeZone;
+use File;
 /**
  * Class FrontendController
  * @package App\Http\Controllers
@@ -31,36 +34,44 @@ class FrontendController extends Controller
 		return view('frontend.uploadentry');
 	}
 
-	public function submitproposal(Request $request)
+	public function submitProposal(Request $request)
 	{
-		
+		set_time_limit(300);
 		$destFolder = env('ZIPFILE_FOLDER');
-		$password = env('ZIPFILE_PASSWORD');
-		$fileName = 'tstfile_'. Input::get('OrganisationName').'_test.txt';
-		Input::file('FileName')->move($destFolder, $fileName); // uploading file to given path
+		$password = env('ZIPFILE_PASSWORD');		
+		$currentDate = new DateTime('now',new DateTimeZone('Asia/Singapore'));		
+		$fileExt = strtoupper(Input::file('UploadedFile')->getClientOriginalExtension());
+		$fileName = strtoupper(sprintf('SFC_%s_%s_FILE.%s',substr(Input::get('OrganisationName'),0,3),$currentDate->format('Ymd_His'),$fileExt));		
+		$zipFilename = $destFolder.str_replace('.'.$fileExt,'.zip',$fileName);		 
 		
 		$sourceFileName = $destFolder.$fileName;
-		$zipFilename = $destFolder.'zipfile_'. Input::get('OrganisationName'). 'test.zip';		
+		if(File::exists($sourceFileName)){			
+			File::delete($sourceFileName);
+		}
 		
-		@system("zip -P $password $zipFilename $sourceFileName -q -j");
-		// $zip = new ZipArchive();
-		// if ($zip->open($zipFilename, ZipArchive::CREATE)!==TRUE) {
-			// return("cannot open zipFilename");
-		// }				
-		// $zip->setPassword(env('ZIPFILE_PASSWORD'));
-		// $zip->addFile($destFolder.$fileName,$fileName);				
-		// $zip->close();
+		Input::file('UploadedFile')->move($destFolder, $fileName); // uploading file to zip folder
 				
-		$data = array( 'filePath' => $zipFilename);		
-		Mail::send('emails.submitproposal', $data, function ($m) use ($data) {
-			$m->from('hello@app.com', 'Your Application');
-			$m->to('eka.suharlim@gmail.com', 'Eka Suharlim')->subject('Your Reminder!');
-			$m->attach($data['filePath']);			 
-		});		
-		
-		unlink($destFolder.$fileName);
-		
-		return 'proposal submited ' .' saved' ;
+		if(File::exists($sourceFileName)){			
+			@system("zip -P $password $zipFilename $sourceFileName -q -j");
+			if(File::exists($zipFilename)){			
+				$data = array( 'filePath' => $zipFilename);		
+				Mail::send('emails.submitproposal', $data, function ($m) use ($data) {
+					$m->from('hello@app.com', 'Your Application');
+					$m->to('eka.suharlim@gmail.com', 'Eka Suharlim')->subject('Your Reminder!');
+					$m->attach($data['filePath']);			 
+				});					
+			}else{
+				return 'File zip upload fails, please try again';
+			}
+			unlink($destFolder.$fileName);			
+		}else{
+			return 'File upload fails, please try again';
+		}
+		return redirect()->route('frontend.submitsuccess');
+	}
+	
+	public function submitSuccess(){
+		return view('frontend.submitsuccess');		
 	}
 
 	/**
